+++ cpp/open3d/core/Tensor.cpp
@@ -42,6 +42,14 @@
 namespace open3d {
 namespace core {
 
+struct Open3DDLManagedTensor {
+    Tensor o3d_tensor_;
+    DLManagedTensor dl_managed_tensor_;
+    static void Deleter(DLManagedTensor* arg) {
+        delete static_cast<Open3DDLManagedTensor*>(arg->manager_ctx);
+    }
+};
+
 /// Tensor assignment lvalue = lvalue, e.g. `tensor_a = tensor_b`
 Tensor& Tensor::operator=(const Tensor& other) & {
     shape_ = other.shape_;
@@ -940,6 +948,161 @@ std::vector<Tensor> Tensor::NonZeroNumpy() const {
 }
 
 Tensor Tensor::NonZero() const { return kernel::NonZero(*this); }
+
+DLManagedTensor* Tensor::ToDLPack() const {
+    DLDeviceType dl_device_type;
+    switch (GetDevice().GetType()) {
+        case Device::DeviceType::CPU:
+            dl_device_type = DLDeviceType::kDLCPU;
+            break;
+        case Device::DeviceType::CUDA:
+            dl_device_type = DLDeviceType::kDLGPU;
+            break;
+        default:
+            utility::LogError("ToDLPack: unsupported device type {}",
+                              GetDevice().ToString());
+    }
+
+    DLContext dl_context;
+    dl_context.device_type = dl_device_type;
+    dl_context.device_id = GetDevice().GetID();
+
+    DLDataType dl_data_type;
+    switch (GetDtype()) {
+        case Dtype::Float32:
+            dl_data_type.code = DLDataTypeCode::kDLFloat;
+            break;
+        case Dtype::Float64:
+            dl_data_type.code = DLDataTypeCode::kDLFloat;
+            break;
+        case Dtype::Int32:
+            dl_data_type.code = DLDataTypeCode::kDLInt;
+            break;
+        case Dtype::Int64:
+            dl_data_type.code = DLDataTypeCode::kDLInt;
+            break;
+        case Dtype::UInt8:
+            dl_data_type.code = DLDataTypeCode::kDLUInt;
+            break;
+        default:
+            utility::LogError("Unsupported data type");
+    }
+    dl_data_type.bits =
+            static_cast<uint8_t>(DtypeUtil::ByteSize(GetDtype()) * 8);
+    dl_data_type.lanes = 1;
+
+    Open3DDLManagedTensor* o3d_dl_tensor(new Open3DDLManagedTensor);
+    o3d_dl_tensor->o3d_tensor_ = *this;
+
+    DLTensor dl_tensor;
+    // Not Blob's data pointer.
+    dl_tensor.data = const_cast<void*>(GetDataPtr());
+    dl_tensor.ctx = dl_context;
+    dl_tensor.ndim = static_cast<int>(GetShape().size());
+    dl_tensor.dtype = dl_data_type;
+    // The shape pointer is alive for the lifetime of Open3DDLManagedTensor.
+    dl_tensor.shape = const_cast<int64_t*>(
+            o3d_dl_tensor->o3d_tensor_.GetShapeRef().data());
+    // The strides pointer is alive for the lifetime of Open3DDLManagedTensor.
+    dl_tensor.strides = const_cast<int64_t*>(
+            o3d_dl_tensor->o3d_tensor_.GetStridesRef().data());
+    dl_tensor.byte_offset = 0;
+
+    o3d_dl_tensor->dl_managed_tensor_.manager_ctx = o3d_dl_tensor;
+    o3d_dl_tensor->dl_managed_tensor_.deleter = &Open3DDLManagedTensor::Deleter;
+    o3d_dl_tensor->dl_managed_tensor_.dl_tensor = dl_tensor;
+
+    return &(o3d_dl_tensor->dl_managed_tensor_);
+}
+
+Tensor Tensor::FromDLPack(const DLManagedTensor* src) {
+    Device device;
+    switch (src->dl_tensor.ctx.device_type) {
+        case DLDeviceType::kDLCPU:
+            device = Device("CPU", src->dl_tensor.ctx.device_id);
+            break;
+        case DLDeviceType::kDLGPU:
+            device = Device("CUDA", src->dl_tensor.ctx.device_id);
+            break;
+        default:
+            utility::LogError("Unsupported device_type {}",
+                              src->dl_tensor.ctx.device_type);
+    }
+
+    Dtype dtype;
+    if (src->dl_tensor.dtype.lanes != 1) {
+        utility::LogError("Only supports lanes == 1, but lanes == {}",
+                          src->dl_tensor.dtype.lanes);
+    }
+    switch (src->dl_tensor.dtype.code) {
+        case DLDataTypeCode::kDLUInt:
+            switch (src->dl_tensor.dtype.bits) {
+                case 8:
+                    dtype = Dtype::UInt8;
+                    break;
+                default:
+                    utility::LogError("Unsupported kDLUInt bits {}",
+                                      src->dl_tensor.dtype.bits);
+            }
+            break;
+        case DLDataTypeCode::kDLInt:
+            switch (src->dl_tensor.dtype.bits) {
+                case 32:
+                    dtype = Dtype::Int32;
+                    break;
+                case 64:
+                    dtype = Dtype::Int64;
+                    break;
+                default:
+                    utility::LogError("Unsupported kDLInt bits {}",
+                                      src->dl_tensor.dtype.bits);
+            }
+            break;
+        case DLDataTypeCode::kDLFloat:
+            switch (src->dl_tensor.dtype.bits) {
+                case 32:
+                    dtype = Dtype::Float32;
+                    break;
+                case 64:
+                    dtype = Dtype::Float64;
+                    break;
+                default:
+                    utility::LogError("Unsupported kDLFloat bits {}",
+                                      src->dl_tensor.dtype.bits);
+            }
+            break;
+        default:
+            utility::LogError("Unsupported dtype code {}",
+                              src->dl_tensor.dtype.code);
+    }
+
+    // Open3D Blob's expects an std::function<void(void*)> deleter.
+    auto deleter = [src](void* dummy) -> void {
+        if (src->deleter != nullptr) {
+            src->deleter(const_cast<DLManagedTensor*>(src));
+        }
+    };
+
+    SizeVector shape(src->dl_tensor.shape,
+                     src->dl_tensor.shape + src->dl_tensor.ndim);
+
+    SizeVector strides;
+    if (src->dl_tensor.strides == nullptr) {
+        strides = Tensor::DefaultStrides(shape);
+    } else {
+        strides = SizeVector(src->dl_tensor.strides,
+                             src->dl_tensor.strides + src->dl_tensor.ndim);
+    }
+
+    auto blob = std::make_shared<Blob>(device, src->dl_tensor.data, deleter);
+
+    // src->dl_tensor.byte_offset is ignored in PyTorch and MXNet, but
+    // according to dlpack.h, we added the offset here.
+    return Tensor(shape, strides,
+                  reinterpret_cast<char*>(blob->GetDataPtr()) +
+                          src->dl_tensor.byte_offset,
+                  dtype, blob);
+}
 
 }  // namespace core
 }  // namespace open3d
